backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3

channels>=4.0
daphne>=4.0

gunicorn>=21.2
uvicorn>=0.23

psycopg2-binary>=2.9

Pillow>=10.0
